Fix deadlock seen when createCaptureSession and encodeImage race.
A deadlock can occur due to the following conditions when we start a
webcam only stream and then switch preview on:
1) Thread 1 : CameraController.setWebcamStreamConfig - holds native
   mSerializationLock, waits on CameraController.mSerializationLock
2) Thread 2 : CameraController.createCaptureSession - holds
   CameraController.mSerializationLock and waits on onConfigured call to
   be handled by mBackgroundHandlerThread
3) Thread 3 : encodeImage is being processed on mBackgroundHandlerThread
   - waits on native mSerializationLock held by thread 1.

Encoding and camera callbacks don't really need to be handled on the
same thread. We create a new thread to handle camera callbacks. Camera
operations are still wrapped in blocking calls to make flow of code
easier since camera can be controlled from both native and java calls.

Bug: 277099495

Test: Start webcam stream and then preview multiple times; don't see ANR

Change-Id: Iec1408c32d593f4a33755c379249d02f8d400db2
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>