Add zoom control function for DeviceAsWebcam
Use ScaleGestureDetector to get the scale factors from the motion events
of the TextureView. Then, calculate the updated zoom ratio by the scale
factor and set it to the working capture session to have the zoom
control functionality.

The zoom ratio setting is cached in CameraController and is retrieved to
set to MotionEventToZoomRatioConverter. So that the last zoom setting
can be restored if the users is in a video call and close/open the
preview activity to adjust the zoom ratio many times. The zoom ratio
setting is reset to be 1.0f when both the webcam and preview streams are
closed.

Bug: 269644311
Test: manual test
Merged-In: I035ae3e48aa28375079587a973b2acd6f9eb353d
Change-Id: I035ae3e48aa28375079587a973b2acd6f9eb353d
(cherry picked from commit c51b46cb7352553fe2f5d7d7729d894d8621ad90)
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>