Rotate webcam stream images to correct orientation
Add new parameter to rotate images when converting to MJpeg format.
Only support to rotate the webcam stream images 180 degrees.

Created a RotationProvider to monitor the orientation changes via
OrientationEventListener and then convert it to the corresponding
rotation degrees value to rotate the images to upright orientation.

Added a CameraInfo class to provide necessary characteristics values of
the working camera.

Bug: 269644311
Test: manual test

Change-Id: I80c274d132d80399fafaeecfbac6a2fd22609768