Make stream rotation UX more intuitive.
The current stream rotation logic flipped the stream at 45 degrees
off the cartesian axes which resulted in a somewhat janky experience.
The stream would also quickly flip between the two rotations when the
device was held at a 45 degree angle and moved slightly.

To remove the jankiness feel, this CL changes the rotation logic to
move trigger points to the cartesian axes which is where the user
would typically expect the rotation to happen. It also adds a debounce
zone near the rotation trigger points so the stream is only flipped when
user intentionally goes past a certain point.

Bug: 286954878
Test: Manually tested that rotation works correctly.
Change-Id: I298f978d18291568886645c219becaaf2817df33