Update camera related fields synchronously when toggling camera.
When the UI requests CameraController to toggle the the camera, it
expects to be able to request zoom ratio (and potentially other camera
related metadata) as soon as the function returns.

The current implementation updates the camera fields and streams
asynchronously which means that the preview might get stale information
until the asynchronous work is scheduled. This CL updates the toggling
logic to update the camera related metadata synchronously, and setup
the camera streams in a background thread. This ensures that the preview
can always query the most up to date camera information.

Test: Tried switching camera 20 times. The UI was always consistent
      with the current webcam state.
Bug: 286954878
Change-Id: I1d231d411fcd1a26c7ba0934b78a012e16f9fe02