Default to last settings when starting webcam.
The general user experience expected to be a somewhat consistent
setup when it comes to using the webcam feature. In those situations,
fiddling with the device to set up the camera choice and zoom at every
connection is fairly annoying.

To make the setup experience smoother, this CL stores last user
preferences of camera and zoom, and restores them on next launch. This
ensures that if the phone is set up with consistent mounts, the webcam
can be used directly with little changes.

Test: Manually verified that the preferences persist between connections
	  and even reboots.
Bug: 287543028
Change-Id: I5a219743c8cd7a412289c0e1cd989f7f2b3f0480