Format RRO README.md
The RRO overlay README is a markdown file, but was not using any of
the features of markdown. This CL updates the README with proper
formatting to (hopefully) make it more easy to read. It also sets up
a pattern that can be repeated for different overridable files that
might be added in the future.

Bug: 290833175
Test: Manually tested that the README render is more user-friendly

Merged-In: I9f9dbf1110f089196ab9a272bda3f9f19c384523
Change-Id: I9f9dbf1110f089196ab9a272bda3f9f19c384523
(cherry picked from commit 469f9eb9a141ee91a7d30fde44e43a4c032081c0)
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>