Allow vendors to custom physical camera zoom ratio range
This is useful when OEMs try to map a logical cameras to one of its underlying physical
camera and the physical camera might have a different supported zoom ratio range from
the logical camera. Overlay of the JSON file can calibrate the supported range of the
zoom function in this case.

Bug: 287541195
Test: manual test
Change-Id: I8b170af37b6d7cb9c143d70605fda2fe97e7d623