Use dp instead of sp for zoom text
The ZoomController has text that act as both status and button.
This zoom text is currently described with in sp units which is
affected by text scaling.

Due to lack of space, and because the exact text in the view is
not critical, this CL switches them to be described by dp
instead. This ensures that the text isn't truncated when text
scaling is increased from the system.

Bug: 296969135
Test: Increased Text Scaling and UI Scaling to max to ensure
      that text doesn't clip.
Change-Id: I0526182aa94f87e33857101e67478f4c71fb4c93