Fix accessbility issues with ZoomController
ZoomController had issues when attempting to navigate with
Talkback, SwitchAccess, or Voice Control. The issues stemmed from
ZoomController relying on relatively low level onTouchListener to
provide an intuitive UX. However Touch Listeners do not play well
with Accessibility Services, which made ZoomController behave weirdly
when using accessibility services.

This CL makes the base ZoomController rely on onClickListener and
onLongPressListener, which play well with the Accessibility service.
When Accessibility Services are not being used, it overlays a transparent
view that has the same onTouch logic as before making the changes
completely invisible when Accessibility Services are not being used.

It also adds a listener to ensure that the ZoomController can toggle
between the two modes if Accessibility Services are enables/disabled
while the preview is running.

Additionally, it makes the following changes to make ZoomController
more accessible:
 - Denote iteration order for accessibility
 - Make zoom TextView's clickable for SwitchAccess
 - Long press zoom buttons in order to get to Zoom SeekBar

Bug: 296959663
Bug: 296960927
Bug: 296967389
Bug: 296973503
Bug: 296980552
Bug: 296980708
Bug: 296982989
Bug: 296983002
Test: Manually tested that Accessibility Services can change zoom values
      and UX when running without Accessibility Services is unchanged.

Change-Id: Ibfc04263828d300be591c6b86a75e579812251c5