Handle display state changes while DeviceAsWebcam Preview is running.
The current implementation of the preview used the max status bar
height seen as a proxy for measuring the display cutout. However,
this failed if the device cutout changed while the Preview Activity
was running. This also unearthed an issue where on a foldable, the
view finder would be incorrectly sized when the display was folded
or unfolded.

This patch makes two changes to fix the issues:
 1. Instead of relying on the status bar height, which returned 0
    when the Activity went fullscreen, we now rely on cutout insets
    which is constant regardless of the Activity state. It is also
    updated when a foldable device changes it folded state.

 2. Listens for layout changes when setting the margin for texture
    view container. After applying the margins we wait for the
    container to be drawn before rescaling the preview to fit the
    newly laid out container.

With these changes, the preview stays consistent for foldables.

Bug: 298705937
Test: Manually tested that the preview isn't clipped or undersized
      on foldables.
Change-Id: Ib2fb66d0c6750ed4ba6fd8465ca79edf09d9cbac