Fix preview logic to incorporate webcam stream size.
The current implementation of calculating preview size
accidentally only considered the display size, instead of both
the display size and the webcam stream size. This would
lead to situation where the preview didn't accurately represent
what was streamed to the host.

This CL fixes the logic to ensure that when a webcam stream is
present, the preview reflects the webcam stream in FoV and aspect
ratio.

Bug: 286954878
Test: Manually tested that Preview changes aspect ratio and FoV
      when not using a 16:9 webcam stream.
Change-Id: I3cd17b9b67db607ba34ea2782d027b8f5f4523ae