Simplify camera availability state callback handling
When we get an onCameraAvailable callback and webcam was streaming, we
should attempt to start streaming again if
a) The available camera id is the one that was streaming earlier
b) The camera id that was streaming earlier had an error / got
   disconnected (since its possible that even though it was 'available'
   the camera sub-system doesn't support opening it along with other
   cameras that some other clients may have opened)

For spurious onCameraAvailable callbacks we'll get an onError callback
when trying to connect, which will be handled by b) the next time an
onCameraAvailable callback occurs.

All operations which require mSerializationLock should also be offloaded
to the service events executor thread.

Bug: 301216738

Test: Start streaming webcam; open GCA; change camera ids on GCA; make
      sure webcam stream is able to eventually start showing the live
      feed again.

Test: Start streaming webcam; kill camera HAL; see that webcam stream
      eventually starts showing live feed again.

Merged-In: I734df42be4130eefb7f35458216f9b30662f2926
Change-Id: I734df42be4130eefb7f35458216f9b30662f2926
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>
(cherry picked from commit 63ce8f798a69ab41acd5a6cfaa1558944fa78a78)