Make DeviceAsWebcam single user.
DeviceAsWebcam is a service which attempts to grab connect to the V4L2
node mounted by the UVC gadget driver. Because UVC gadget driver has
no concept of "users", the DeviceAsWebcam service needs to run as a
singleton across the whole device.

This CL marks all components of the service with "android:singleUser"
to ensure that there is only one instance of DeviceAsWebcam process
running at one time.

Bug: 302948147
Test: Manually tested that only one service process is running
      even when multiple users are enrolled to the device.
Change-Id: Ia535d4622b7d673eaa5602aeb3cf2c324703d183