Add High Quality option
This CL adds a 'High Quality' toggle that the users can
pick if they want to trade performance for quality. This
option drops all RRO and StreamUseCase optimizations to use
the cameras a regular video camera application would, maximizing
quality.

Follow up CLs will add a warning to let the users know of
potential long term damage to the device if this mode is
used for an extended period of time.

Bug: 313179507
Test: Manually verified that the button works
Change-Id: Ifebc12d1b783c4db499a0f97b230e6a35a4a7ad0