Fix largescreen UI and rotation handling
DeviceAsWebcam's preview activity was forced to portrait
which led to letter/pillarboxing in large screen devices and
disallowed system rotation.

With this CL the activity is fixed to the device's natural
orientation programmatically, allowing the system to rotate
according to user preferences. This also disables
letter/pillarboxing on large screen devices and uses the full
screen for consistent behavior on all devices and configurations.

This also acts as a starting point to support different
foldable configurations.

Bug: 306204649
Test: Manually tested on candybar phones and foldables.
Change-Id: I2239be3de0ace8dae95583d70cc5870adeddb73d