Add warning for High Quality Mode
High Quality Mode was added to DeviceAsWebcam in a previous CL which
disables power optimizations to boost camera quality.

This CL adds warning dialog that pops up when the user enables high
quality mode to ensure that the users are aware of potential long term
effects on the battery when High Quality Mode is used.

Bug: 313179507
Bug: 320857464
Test: Manually tested that the dialog works.
Change-Id: Id674928204365166bb9811c03d4d61165479886b