Make Camera Selector a Dialog
The current implementation of Camera Selector created a
dialog by overlaying a View on the main Activity. However this
meant hardcoding a lot of values and making assumptions about the
screen it was being drawn on. For example, it makes the assumption
that the selector will always be shown at a 90 degree angle to the
device's natural orientation. This caused some maintenance and
testing overhead when adding support for large screen devices.

This patch moves the CameraSelector into an AlertDialog that can be
drawn natively by the system and doesn't need any hardcoding or special
handling when dealing with large screen and variable screen devices.

As a bonus, it also fixes accessibility issues stemming from hardcoding
the dialog dimensions.

Bug: 306325960
Test: Manually Tested that the dialog works
Change-Id: I6a6d7a94cd645e30473df8eb072eb9e47df10858