Handle all cutout insets
WindowManager was updated to enforce that apps cannot use
android:windowLayoutInDisplayCutoutMode = "shortEdges". The
only allowed value is now "always", which means applications
now have to handle cutouts on all edges, instead of just
the shorter edges (translating to top and bottom for most
devices).

This CL ensures that DeviceAsWebcamPreview handles all cutouts
instead of just top and bottom ones as it was doing before.

Bug: 331223541
Test: Manually Verified that the Preview works and the view stays
      unchanged through rotations.
Change-Id: I2c001f1516a18bc91a0ad3e8645f0b02819d33c9